% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_analysis.R
\name{tidy_analysis}
\alias{tidy_analysis}
\title{Create an output format for tidy analysis from pdf_document}
\usage{
tidy_analysis(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link{pdf_document}}}
}
\value{
This function returns an output format to generate a pdf document.
}
\description{
Create an output format for tidy analysis from pdf_document
}
\note{
This function is not designed to be run from the console, but from the YAML header of the R Markdown document.
}
